end and start are stages now
