+++ src/main/java/fsm/process/ChooseSyntax.java
@@ -1,20 +1,20 @@
-package fsm;
+package fsm.process;
 
+import fsm.Guard;
 import lombok.RequiredArgsConstructor;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Consumer;
-import java.util.function.Predicate;
 
 public class ChooseSyntax {
     final List<Option> options = new ArrayList<>();
 
-    ChooseSyntax when(Guard guard, Consumer<Builder> config) {
+    ChooseSyntax when(Guard guard, Consumer<Process> config) {
         options.add(new Option(guard, config));
         return this;
     }
-    ChooseSyntax otherwise(Consumer<Builder> config) {
+    ChooseSyntax otherwise(Consumer<Process> config) {
         return when(Guard.ALLOW, config);
     }
     static ChooseSyntax choose() {
@@ -24,6 +24,6 @@ public class ChooseSyntax {
     @RequiredArgsConstructor
     static class Option {
         final Guard guard;
-        final Consumer<Builder> consumer;
+        final Consumer<Process> consumer;
     }
 }
